moved all iterators to a separate file
